:sparkles: feat/(#41, S11P31D207-176, S11P31D207-177): Vital 도메인 컨트롤러 및 서비스 작업

+++ backend/mongddang/src/main/java/com/onetwo/mongddang/domain/medication/model/MedicationTime.java
@@ -27,10 +27,10 @@ public class MedicationTime {
     private Boolean isFast;
 
     @Column(name="min_glucose")
-    private Long minGlucose;
+    private int minGlucose;
 
     @Column(name="max_glucose")
-    private Long maxGlucose;
+    private int maxGlucose;
 
     @NotNull
     private Long volume;
